+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* Support migrating projects using Setuptools with `setup.cfg`
 * [poetry] Do not set optional groups as default ones ([#299](https://github.com/mkniewallner/migrate-to-uv/pull/299))
 * Indicate support for Python 3.14